# api.yaml
swagger: "2.0.0"
info:
  title: Test API Documentation
  description: Api Documentation for Test App Cyber Flow
  version: 1.0.0
  contact:
    email: choudharysidharth082000@gmail.com
  servers:
    - url: "http://localhost:3000"
      description: "Local Server Running"
    - url: "https://testcyberflow.herokuapp.com"
      description: "Main Server"
schemes: [http, https]
paths:     
...
  /v1/profile/editProfileImage/{userID}:
    put:
      tags:
        - name: Profile
      summary: PRofile Image Update
      description: Will Update Profile Image of the user
      consumes:
        multipart/form-data
      produces:
        application/json
      parameters:
        - name: userID
          in: path
          type: string 
          required: true
        - in: formData
          name: profileImage
          type: file 
          required: true
      responses:
        200:
          description: OK
  /v1/profile/updateProfileSection/{userID}:
    put:
      tags:
        - name: Profile
      summary: Update Profile
      description: APi will update the profileImage
      consumes:
        application/json
      produces:
        application/json
      parameters:
        - name: userID 
          in: path
          type: string
          required: true
        - in: body
          name: profile 
          schema:
            type: object
            properties:
              name: 
                type: string
              mobileNumber:
                type: string
              about:
                type: string
      responses:
        200:
          description: OK
...
  /v1/profile/getProfileDetails/{userID}:
    get:
      tags:
        - name: Profile
      summary: Profile Details
      description: Getting Basic Info Of Profile
      parameters:
        - name: userID 
          in: path
          type: string
          required: true
      responses:
        200:
          description: OK
...
  /v1/profile/profileUser/{userID}:
    post:
      tags:
        - name: Profile
      summary: Profile Database
      description: APi will be used to update profile Database
      consumes:
        multipart/form-data
      produces:
        application/json
      parameters:
        - name: userID 
          in: path
          type: string
          required: true
        - in: formData
          name: profileImage
          type: file
          description: Profile Photo of the user
        - in: formData
          name: name
          type: string 
          description: name of the customer
        - in: formData
          name: mobileNumber
          type: string 
          description: Phone Number of the customer
        - in: formData
          name: about
          type: string 
          description: about of the customer
      responses:
        200:
          description: OK
...